# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TyxNet API
  version: 0.1.0
  description: Experimental management, enrollment and device-control API.
servers:
  - url: https://vpn.example.com:8443
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  schemas:
    Error:
      type: object
      properties: {error: {type: object}}
    Login:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string, format: password}}
    Enrollment:
      type: object
      required: [Token, Name, public_key]
      properties:
        Token: {type: string}
        Name: {type: string}
        OS: {type: string}
        Arch: {type: string}
        Version: {type: string}
        public_key: {type: string, format: byte}
paths:
  /healthz:
    get: {responses: {'200': {description: Healthy}}}
  /api/v1/auth/login:
    post:
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Login'}}}}
      responses: {'200': {description: Bearer session}, '401': {description: Invalid credentials}}
  /api/v1/auth/logout:
    post: {security: [{bearerAuth: []}], responses: {'204': {description: Logged out}}}
  /api/v1/auth/me:
    get: {security: [{bearerAuth: []}], responses: {'200': {description: Current user}}}
  /api/v1/enroll:
    post:
      description: Consumes an enrollment token and records only the device public key.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Enrollment'}}}}
      responses: {'201': {description: Enrolled device}, '403': {description: Rejected}}
  /control/v1/challenge:
    post: {description: Issues a one-use 30-second device challenge, responses: {'200': {description: Challenge}}}
  /control/v1/connect:
    post: {description: Verifies Ed25519 proof and opens an SSE control stream, responses: {'200': {description: Event stream}, '403': {description: Invalid proof}}}
  /api/v1/devices:
    get: {security: [{bearerAuth: []}], responses: {'200': {description: Devices}}}
  /api/v1/devices/{id}:
    delete:
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses: {'204': {description: Revoked}}
  /api/v1/devices/{id}/{action}:
    post:
      description: Queues restart, shutdown or disconnect. No arbitrary command text is accepted.
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: id, required: true, schema: {type: string}}
        - {in: path, name: action, required: true, schema: {type: string, enum: [restart, shutdown, disconnect]}}
      responses: {'202': {description: Queued command}}
  /api/v1/users:
    get: {security: [{bearerAuth: []}], responses: {'200': {description: Users}}}
    post: {security: [{bearerAuth: []}], responses: {'201': {description: User created}}}
  /api/v1/users/{id}:
    delete:
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses: {'204': {description: Deleted}, '409': {description: User owns devices}}
  /api/v1/tokens:
    get: {security: [{bearerAuth: []}], responses: {'200': {description: Token metadata; never raw token values}}}
    post: {security: [{bearerAuth: []}], responses: {'201': {description: Token value returned once}}}
  /api/v1/tokens/{id}:
    delete:
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses: {'204': {description: Revoked}}
  /api/v1/server/status:
    get: {security: [{bearerAuth: []}], responses: {'200': {description: Status counters}}}